requirements.txt
import pandas
mysql-connector==2.2.9
numpy==1.17.2
pandas==0.25.1
prettytable==0.7.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
six==1.12.0
